feat: Add route to get basic information of users in a workspace
This commit adds a new route `GET /:workspaceId/members` to the user controller. This route allows authenticated users to get the basic information of all the users in a specific workspace. It checks the user's authentication, workspace existence, and user's rights before returning the information. If successful, it responds with a JSON array containing the user data. If there are any errors, it logs them and responds with an appropriate error message.