feat: Add routes to retrieve today and tomorrow tasks for a user
This commit adds two new routes, `/tasks/today` and `/tasks/tomorrow`, to retrieve all tasks that are due today and tomorrow respectively for a specific user. The tasks are filtered based on the user ID and status, excluding archived tasks. In case of any error while retrieving the tasks, a 500 error response with an error message is returned.